chore(dgeni): refactor processors
refactor processor to meaningful name
extract apiRoute from navMenu to a separate processor
